requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0
matplotlib>=3.7.0